[ALPS02845381] M4U:M4U secure driver
[Detail]
MT6799 M4U secure driver suppor 64 bit PA.

MTK-Commit-Id: 20f92864c88da22638c2acd84cb17e2d30c9da1d

CR-Id: ALPS02845381
Feature: Play Ready

+++ drivers/misc/mediatek/include/mt-plat/mt6799/include/trustzone/m4u/tz_m4u.h
@@ -88,14 +88,14 @@ typedef struct {
 	int port;
 	unsigned int mva;
 	unsigned int size;
-	unsigned int pa;
+	unsigned long long pa;
 } m4u_buf_param_t;
 
 typedef struct {
 	unsigned long long nonsec_pt_pa;
 	int l2_en;
 	unsigned int sec_pt_pa;
-	unsigned int sec_pa_start;
+	unsigned long long sec_pa_start;
 	unsigned int sec_pa_size;
 } m4u_init_param_t;
